steward metadata and balances managers
