+++ packages/server/src/services/agents/steward/balances/types.ts
@@ -1,0 +1,12 @@
+import { NetworkURN } from '@/lib.js'
+import { SubstrateIngressConsumer } from '@/services/networking/substrate/ingress/types.js'
+import { Subscription } from 'rxjs'
+
+export type BalancesSubscriptionMapper = (
+  chainId: NetworkURN,
+  ingress: SubstrateIngressConsumer,
+  enqueue: EnqueueJob,
+) => Subscription[]
+
+export type BalanceUpdateJob = () => Promise<void>
+export type EnqueueJob = (key: string, job: BalanceUpdateJob) => void
