+++ packages/server/src/services/agents/steward/metadata/queries/location/handler.spec.ts
@@ -6,7 +6,7 @@ import { of } from 'rxjs'
 import { RuntimeApiContext } from '@/services/networking/substrate/context.js'
 import { SubstrateIngressConsumer } from '@/services/networking/substrate/ingress/types.js'
 import { SubstrateApiContext } from '@/services/networking/substrate/types.js'
-import { AssetMetadata } from '../../types.js'
+import { AssetMetadata } from '../../../types.js'
 import { LocationQueryHandler } from './handler.js'
 
 vi.mock('../../mappers.js', () => {
