+++ packages/server/src/services/agents/steward/metadata/queries/location/util.spec.ts
